tweaked issues with the profile alignment

+++ modules/local/mafft/main.nf
@@ -55,16 +55,16 @@ process MAFFT_ADD_PROFILE {
     label "mafft"
 
     input:
-    tuple val(grouping_id), file(input_files), val(metadatas)
+    tuple val(grouping_id), val(input_files), val(metadatas)
 
     output:
     tuple path("*.fasta"), val(grouping_id), emit: profile_alignment_tuple
 
     script:
-
+    println(input_files)
     file_one = input_files[0]
     other_files = input_files[1..-1]
-    def output_file = "${grouping_id}.profile_aligned.fasta"
+    def output_file = "CAP${grouping_id}.profile_aligned.fasta"
 
     bash_other_files = "(" + other_files.join(" ") + ")"
 
@@ -75,9 +75,14 @@ process MAFFT_ADD_PROFILE {
 
     for file in \${file_arr[@]};
     do
-        mafft --thread -1 --add \$file ${output_file} > temp_${output_file};
-        rm ${output_file};
-        mv temp_${output_file} ${output_file};
+        if [[ -s \$file ]]
+        then
+            mafft --thread -1 --add \$file ${output_file} > temp_${output_file};
+            rm ${output_file};
+            mv temp_${output_file} ${output_file};
+        else
+            echo "The file \$file is empty" 1>&2;
+        fi
     done
     """
 }
